BUG: Code did not build under ITK 3.2
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=4065
git-svn-id: http://svn.slicer.org/Slicer4/trunk@4065 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/CLI/LabelMapSmoothing.cxx
@@ -132,7 +132,7 @@ int main( int argc, char * argv[] )
   for( unsigned int i = 0; i < ImageDimension; i++ )
     {
     boundingBox[2*i] = vnl_math_max(0,(int)(boundingBox[2*i]-boundingBoxPadding));
-    boundingBox[2*i+1] = vnl_math_min(imageSize[i]-1,boundingBox[2*i+1]+boundingBoxPadding);
+    boundingBox[2*i+1] = vnl_math_min((int)(imageSize[i]-1),(int)(boundingBox[2*i+1]+boundingBoxPadding));
     regionIndex[i] = boundingBox[2*i];
     regionSize[i] = boundingBox[2*i+1]-boundingBox[2*i]+1;
     }
